+++ Application/app/src/main/java/com/istudio/distancetracker/di/NotificationModule.kt
@@ -7,7 +7,7 @@ import android.content.Context
 import android.content.Intent
 import android.os.Build
 import androidx.core.app.NotificationCompat
-import com.istudio.distancetracker.MainActivity
+import com.istudio.distancetracker.ui.MainActivity
 import com.istudio.distancetracker.R
 import com.istudio.distancetracker.utils.Constants.ACTION_NAVIGATE_TO_MAPS_FRAGMENT
 import com.istudio.distancetracker.utils.Constants.NOTIFICATION_CHANNEL_ID
