+++ Application/app/src/main/java/com/istudio/distancetracker/ui/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.istudio.distancetracker
+package com.istudio.distancetracker.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
+import com.istudio.distancetracker.R
 import com.istudio.distancetracker.utils.Permissions
 import dagger.hilt.android.AndroidEntryPoint
