+++ Application/app/src/main/res/layout/fragment_permission.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PermissionFragment">
+    tools:context=".ui.permission.PermissionFragment">
 
     <ImageView
         android:id="@+id/welcomeImageView"
